Update check-configurations.mdx (#896)
Removed the mention of "years" in acceptable validation period entries. We may hold data further back than the obligatory 90days for enterprise customers, but we can't guarantee it. As we only promise up to 90 days, we don't need to mention years here and prompt questions from customers about a longer retention period.